pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "bambu-cli"
dynamic = ["version"]
description = "CLI tool for use with Bambu Printers"
authors = [{ name = "Andrew Todd", email = "thegeektechworkshop@gmail.com" }]
dependencies = [
    "paho-mqtt>=2.1.0",
    "ftputil>=5.1.0",
    "enlighten>=1.12.4",
    "sshkeyboard>=2.3.1",
    "rich>=13.9.4",
    "cloudscraper>=1.2.71",
    "certifi>=2024.8.30",
    "ngrok>=1.4.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "GPL3" }

[project.scripts]
bambu = "bambucli.cli:main"


[tool.pdm]
distribution = true

[tool.pdm.version]
source = "scm"
tag_regex = "v(.+)"

[tool.pdm.options]
add = ["--no-self"]
install = ["--no-self"]
remove = ["--no-self"]

[tool.pdm.scripts]
_.working_dir = "src"
bambu = { call = "bambucli.cli:main" }
